requirements.txt
python-telegram-bot==20.0
python-dotenv==0.21.0
APScheduler==3.9.1
httpx==0.23.0